papers/TIARA/requirements.txt
jsonpickle
ftfy
pytorch_pretrained_bert
pytorch_transformers
word2number
conllu
numpy
datasets
transformers==4.17.0
parsimonious
unidecode
sklearn
inflection
strsimpy
stanza
spacy
overrides==3.0
editdistance
nltk
networkx
sacrebleu
SPARQLWrapper
tensorboard
tensorboardx==2.2
tqdm
stanfordnlp
scikit-learn